+++ TryTestWeb/Models/ModelosSistemaContable/ContabilidadAuxiliares/ContabilidadAuxViewModels/EstadoCuentasCorrientesViewModel.cs
@@ -12,7 +12,7 @@ public class EstadoCuentasCorrientesViewModel
     public string NombrePrestador { get; set; }
     public CuentaContableModel CuentaContable { get; set; }
     public DateTime Fecha { get; set; }
-    public int Folio { get; set; }
+    public long Folio { get; set; }
     public string Comprobante { get; set; }
     public TipoDte Documento { get; set; }
     public DateTime Vencim { get; set; }
